// src/generated/gladia/schema/streamingRequest.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Gladia Control API
 * OpenAPI spec version: 1.0
 */
import type { StreamingSupportedEncodingEnum } from "./streamingSupportedEncodingEnum"
import type { StreamingSupportedBitDepthEnum } from "./streamingSupportedBitDepthEnum"
import type { StreamingSupportedSampleRateEnum } from "./streamingSupportedSampleRateEnum"
import type { StreamingRequestCustomMetadata } from "./streamingRequestCustomMetadata"
import type { StreamingSupportedModels } from "./streamingSupportedModels"
import type { LanguageConfig } from "./languageConfig"
import type { PreProcessingConfig } from "./preProcessingConfig"
import type { RealtimeProcessingConfig } from "./realtimeProcessingConfig"
import type { PostProcessingConfig } from "./postProcessingConfig"
import type { MessagesConfig } from "./messagesConfig"
import type { CallbackConfig } from "./callbackConfig"

export interface StreamingRequest {
  /** The encoding format of the audio stream. Supported formats: 
- PCM: 8, 16, 24, and 32 bits 
- A-law: 8 bits 
- μ-law: 8 bits 

Note: No need to add WAV headers to raw audio as the API supports both formats. */
  encoding?: StreamingSupportedEncodingEnum
  /** The bit depth of the audio stream */
  bit_depth?: StreamingSupportedBitDepthEnum
  /** The sample rate of the audio stream */
  sample_rate?: StreamingSupportedSampleRateEnum
  /**
   * The number of channels of the audio stream
   * @minimum 1
   * @maximum 8
   */
  channels?: number
  /** Custom metadata you can attach to this live transcription */
  custom_metadata?: StreamingRequestCustomMetadata
  /** The model used to process the audio. "solaria-1" is used by default. */
  model?: StreamingSupportedModels
  /**
   * The endpointing duration in seconds. Endpointing is the duration of silence which will cause an utterance to be considered as finished
   * @minimum 0.01
   * @maximum 10
   */
  endpointing?: number
  /**
   * The maximum duration in seconds without endpointing. If endpointing is not detected after this duration, current utterance will be considered as finished
   * @minimum 5
   * @maximum 60
   */
  maximum_duration_without_endpointing?: number
  /** Specify the language configuration */
  language_config?: LanguageConfig
  /** Specify the pre-processing configuration */
  pre_processing?: PreProcessingConfig
  /** Specify the realtime processing configuration */
  realtime_processing?: RealtimeProcessingConfig
  /** Specify the post-processing configuration */
  post_processing?: PostProcessingConfig
  /** Specify the websocket messages configuration */
  messages_config?: MessagesConfig
  /** If true, messages will be sent to configured url. */
  callback?: boolean
  /** Specify the callback configuration */
  callback_config?: CallbackConfig
}